pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "essay-pipeline"
version = "0.1.0"
description = "Frontmatter validator and essay indexer for ORGAN-V public-process"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
essay-validate = "src.validator:main"
essay-index = "src.indexer:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
